# API-Users/src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"typeSigup": {
				"properties": {
					"surName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"role": {
						"type": "string"
					}
				},
				"required": [
					"surName",
					"lastName",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"typeVerification": {
				"properties": {
					"email": {
						"type": "string"
					},
					"code": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
			"typeLogin": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"typeUsers": {
				"properties": {
					"surName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"given_name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone_number": {
						"type": "string"
					},
					"dob": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"martital_status": {
						"type": "string"
					},
					"role": {
						"type": "string"
					}
				},
				"required": [
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"educational": {
				"properties": {
					"academic": {
						"type": "string"
					},
					"school": {
						"type": "string"
					},
					"degree": {
						"type": "string"
					},
					"major": {
						"type": "string"
					},
					"end_date": {
						"type": "string"
					},
					"start_date": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"experience": {
				"properties": {
					"years_of_experience": {
						"type": "string"
					},
					"company_Name": {
						"type": "string"
					},
					"position": {
						"type": "string"
					},
					"start_date": {
						"type": "string"
					},
					"end_date": {
						"type": "string"
					},
					"job_description": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"reference": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone_number": {
						"type": "string"
					},
					"company": {
						"type": "string"
					},
					"position": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ProfileCompletion": {
				"properties": {
					"basic": {
						"type": "number",
						"format": "double"
					},
					"experience": {
						"type": "number",
						"format": "double"
					},
					"education": {
						"type": "number",
						"format": "double"
					},
					"reference": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"basic",
					"experience",
					"education",
					"reference"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutUser": {
				"properties": {
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"dob": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"phone_number": {
						"type": "string"
					},
					"martital_status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "user-service-auth",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
		"/v1/user/hello/jest": {
			"get": {
				"operationId": "HelloJest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/v1/user": {
			"post": {
				"operationId": "PostUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/typeSigup"
							}
						}
					}
				}
			}
		},
		"/v1/user/verify": {
			"post": {
				"operationId": "VerifyUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/typeVerification"
							}
						}
					}
				}
			}
		},
		"/v1/user/login": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"idUser": {}
									},
									"required": [
										"idUser"
									],
									"type": "object",
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/typeLogin"
							}
						}
					}
				}
			}
		},
		"/v1/user/{id}": {
			"put": {
				"operationId": "UpdataInformation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/typeUsers"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/typeUsers"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"pf": {
											"type": "string"
										},
										"username": {
											"type": "string"
										}
									},
									"required": [
										"pf",
										"username"
									],
									"type": "object",
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"GetUser"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/user/education/{id}": {
			"put": {
				"operationId": "Educational",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/educational"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetEducation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/educational"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"GetUser"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/user/experince/{id}": {
			"put": {
				"operationId": "Experience",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/experience"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/experience"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetExp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/experience"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"GetUser"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/user/reference/{id}": {
			"put": {
				"operationId": "Reference",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/reference"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/reference"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetReference",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/reference"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"GetUser"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/user/cv/{id}": {
			"put": {
				"operationId": "UpdateCv",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file_path": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"file_path"
								]
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetCvs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"GetUser"
				],
				"security": [],
				"parameters": []
			}
		},
		"/v1/user/cv/{id}/{index}": {
			"delete": {
				"operationId": "DeleteCv",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "index",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/user/photo/{id}": {
			"post": {
				"operationId": "UpdatePhoto",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"photo": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"photo"
								]
							}
						}
					}
				}
			}
		},
		"/v1/user/pfComplete/{id}": {
			"get": {
				"operationId": "GetProfileComplete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ProfileCompletion"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"GetUser"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/user/bacis/{id}": {
			"get": {
				"operationId": "GetBasic",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/PutUser"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"GetUser"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/generate/token": {
			"get": {
				"operationId": "GenerateToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"generate token"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}